support fp8_e4m3 matmul quantization (merge request !333)

+++ src/ksana_llm/layers/ascend/gptq_matmul_layer.cpp
@@ -12,7 +12,7 @@ Status GPTQMatMulLayer<T, WT>::Init(const std::vector<std::any>& parameters, std
 }
 
 template <typename T, DataType WT>
-int GPTQMatMulLayer<T, WT>::GetWorkSpaceSize() {
+size_t GPTQMatMulLayer<T, WT>::GetWorkSpaceSize() {
   return 0;
 }
