+++ src/ksana_llm/layers/gptq_matmul_layer.h
@@ -12,7 +12,7 @@ class GPTQMatMulLayer : public BaseLayer {
  public:
   virtual Status Init(const std::vector<std::any>& parameters, std::shared_ptr<Context> context, int rank) override;
 
-  virtual int GetWorkSpaceSize() override;
+  virtual size_t GetWorkSpaceSize() override;
 
   virtual Status Forward(const std::vector<Tensor>& input_tensors, std::vector<Tensor>& output_tensors) override;
